refactor: 위시리스트 구현을 위해 엔티티 관계 재정립 (ios-h/airbnb#14)
- Customer 에서 Wish를 참조해서 가져오는 것으로 개발
